+++ src/apis/index.ts
@@ -1,4 +1,29 @@
-export * from './state'
+export {
+  del,
+  isReactive,
+  isRef,
+  isRaw,
+  markRaw,
+  reactive,
+  ref,
+  customRef,
+  Ref,
+  set,
+  shallowReactive,
+  shallowRef,
+  toRaw,
+  toRef,
+  toRefs,
+  triggerRef,
+  unref,
+  UnwrapRef,
+  isReadonly,
+  shallowReadonly,
+  proxyRefs,
+  ShallowUnwrapRef,
+  readonly,
+  DeepReadonly,
+} from '../reactivity'
 export * from './lifecycle'
 export * from './watch'
 export * from './computed'
@@ -7,3 +32,4 @@ export { useCSSModule } from './useCssModule'
 export { createApp } from './createApp'
 export { nextTick } from './nextTick'
 export { createElement as h } from './createElement'
+export { warn } from './warn'