+++ src/component/componentOptions.ts
@@ -1,16 +1,31 @@
+import Vue, { VNode, ComponentOptions as Vue2ComponentOptions } from 'vue'
 import { Data } from './common'
 import { ComponentPropsOptions, ExtractPropTypes } from './componentProps'
-import { VNode } from 'vue'
 import { ComponentInstance, ComponentRenderProxy } from './componentProxy'
 
-import { ComponentOptions as Vue2ComponentOptions } from 'vue'
-
 export interface SetupContext {
-  readonly attrs: Record<string, string>
-  readonly slots: { [key: string]: (...args: any[]) => VNode[] }
+  readonly attrs: Data
+  readonly slots: Readonly<{ [key in string]?: (...args: any[]) => VNode[] }>
+
+  /**
+   * @deprecated not avaliable in Vue 3
+   */
   readonly parent: ComponentInstance | null
+
+  /**
+   * @deprecated not avaliable in Vue 3
+   */
   readonly root: ComponentInstance
-  readonly listeners: { [key: string]: Function }
+
+  /**
+   * @deprecated not avaliable in Vue 3
+   */
+  readonly listeners: { [key in string]?: Function }
+
+  /**
+   * @deprecated not avaliable in Vue 3
+   */
+  readonly refs: { [key: string]: Vue | Element | Vue[] | Element[] }
 
   emit(event: string, ...args: any[]): void
 }
@@ -43,12 +58,15 @@ interface ComponentOptionsBase<
   D = Data,
   C extends ComputedOptions = {},
   M extends MethodOptions = {}
->
-  extends Omit<
+> extends Omit<
     Vue2ComponentOptions<Vue, D, M, C, Props>,
     'data' | 'computed' | 'method' | 'setup' | 'props'
   > {
-  data?: (this: Props, vm: Props) => D
+  // allow any custom options
+  [key: string]: any
+
+  // rewrite options api types
+  data?: (this: Props & Vue, vm: Props) => D
   computed?: C
   methods?: M
 }
@@ -70,6 +88,7 @@ export type ComponentOptionsWithProps<
   Props = ExtractPropTypes<PropsOptions>
 > = ComponentOptionsBase<Props, D, C, M> & {
   props?: PropsOptions
+  emits?: string[] | Record<string, null | ((emitData: any) => boolean)>
   setup?: SetupFunction<Props, RawBindings>
 } & ThisType<ComponentRenderProxy<Props, RawBindings, D, C, M>>
 
@@ -82,6 +101,7 @@ export type ComponentOptionsWithArrayProps<
   Props = Readonly<{ [key in PropNames]?: any }>
 > = ComponentOptionsBase<Props, D, C, M> & {
   props?: PropNames[]
+  emits?: string[] | Record<string, null | ((emitData: any) => boolean)>
   setup?: SetupFunction<Props, RawBindings>
 } & ThisType<ComponentRenderProxy<Props, RawBindings, D, C, M>>
 
@@ -93,6 +113,7 @@ export type ComponentOptionsWithoutProps<
   M extends MethodOptions = {}
 > = ComponentOptionsBase<Props, D, C, M> & {
   props?: undefined
+  emits?: string[] | Record<string, null | ((emitData: any) => boolean)>
   setup?: SetupFunction<Props, RawBindings>
 } & ThisType<ComponentRenderProxy<Props, RawBindings, D, C, M>>