+++ src/component/componentProxy.ts
@@ -1,4 +1,4 @@
-import { ExtractPropTypes } from './componentProps'
+import { ExtractDefaultPropTypes, ExtractPropTypes } from './componentProps'
 import { ShallowUnwrapRef } from '..'
 import { Data } from './common'
 
@@ -22,10 +22,16 @@ export type ComponentRenderProxy<
   D = {}, // return from data()
   C extends ComputedOptions = {},
   M extends MethodOptions = {},
-  PublicProps = P
+  PublicProps = P,
+  Defaults = {},
+  MakeDefaultsOptional extends boolean = false
 > = {
   $data: D
-  $props: Readonly<P & PublicProps>
+  $props: Readonly<
+    MakeDefaultsOptional extends true
+      ? Partial<Defaults> & Omit<P & PublicProps, keyof Defaults>
+      : P & PublicProps
+  >
   $attrs: Data
 } & Readonly<P> &
   ShallowUnwrapRef<B> &
@@ -35,11 +41,22 @@ export type ComponentRenderProxy<
   Omit<Vue, '$data' | '$props' | '$attrs'>
 
 // for Vetur and TSX support
-type VueConstructorProxy<PropsOptions, RawBindings> = VueConstructor & {
+type VueConstructorProxy<
+  PropsOptions,
+  RawBindings,
+  Data,
+  Computed extends ComputedOptions,
+  Methods extends MethodOptions
+> = VueConstructor & {
   new (...args: any[]): ComponentRenderProxy<
     ExtractPropTypes<PropsOptions>,
     ShallowUnwrapRef<RawBindings>,
-    ExtractPropTypes<PropsOptions, false>
+    Data,
+    Computed,
+    Methods,
+    ExtractPropTypes<PropsOptions>,
+    ExtractDefaultPropTypes<PropsOptions>,
+    true
   >
 }
 
@@ -51,14 +68,14 @@ export type VueProxy<
   PropsOptions,
   RawBindings,
   Data = DefaultData<Vue>,
-  Computed = DefaultComputed,
-  Methods = DefaultMethods<Vue>
+  Computed extends ComputedOptions = DefaultComputed,
+  Methods extends MethodOptions = DefaultMethods<Vue>
 > = Vue2ComponentOptions<
   Vue,
   ShallowUnwrapRef<RawBindings> & Data,
   Methods,
   Computed,
   PropsOptions,
-  ExtractPropTypes<PropsOptions, false>
+  ExtractPropTypes<PropsOptions>
 > &
-  VueConstructorProxy<PropsOptions, RawBindings>
+  VueConstructorProxy<PropsOptions, RawBindings, Data, Computed, Methods>